requirements.txt
numpy>=1.19.0
matplotlib>=3.3.0
python-dotenv>=0.15.0
opencv-python>=4.5.0
pillow>=10.0.0
# Python.NET is required by gta5_modules (CodeWalker.Core.dll integration) on Linux + Windows.
pythonnet>=3.0.0